# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/genres": {
      "post": {
        "operationId": "Genres_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreListResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Genres"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "create",
        "description": "Requires role: admin."
      },
      "get": {
        "operationId": "Genres_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreListResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Genres"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "find All"
      }
    },
    "/genres/{id}": {
      "get": {
        "operationId": "Genres_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Genres"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "find One"
      },
      "patch": {
        "operationId": "Genres_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGenreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Genres"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "update",
        "description": "Requires role: admin."
      },
      "delete": {
        "operationId": "Genres_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Genres"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "remove",
        "description": "Requires role: admin."
      }
    },
    "/movies": {
      "get": {
        "operationId": "Movies_getMovies",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "format": "int32",
              "nullable": true,
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "format": "int32",
              "nullable": true,
              "default": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieListResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Movies"
        ],
        "security": [],
        "summary": "get Movies"
      },
      "post": {
        "operationId": "Movies_createMovie",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMovieDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedMovieResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Movies"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "create Movie"
      }
    },
    "/movies/{id}": {
      "get": {
        "operationId": "Movies_getMovieById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Movies"
        ],
        "security": [],
        "summary": "get Movie By Id"
      },
      "put": {
        "operationId": "Movies_updateMovieById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMovieDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedMovieResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Movies"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "update Movie By Id"
      }
    },
    "/movies/tmdb/search": {
      "get": {
        "operationId": "Movies_searchMoviesFromTmdb",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "format": "int32",
              "nullable": true,
              "default": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbSearchResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Movies"
        ],
        "security": [],
        "summary": "search Movies From Tmdb"
      }
    },
    "/movies/tmdb/genres": {
      "get": {
        "operationId": "Movies_getTmdbGenres",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbGenresResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Movies"
        ],
        "security": [],
        "summary": "get Tmdb Genres"
      }
    },
    "/movies/tmdb/{movieId}": {
      "get": {
        "operationId": "Movies_getTmdbMovieDetails",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbMovieResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Movies"
        ],
        "security": [],
        "summary": "get Tmdb Movie Details"
      }
    },
...
    "/ratings/movie/{id}": {
      "get": {
        "operationId": "Ratings_getOneRating",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatedMovieResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Ratings"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get One Rating"
      }
    },
...
  },
  "info": {
    "title": "Paradiso API",
    "description": "Movie sharing, rooms, ratings and invitations. Successful responses include apiVersion. Use an access token with Authorize; refresh tokens are only for /auth/refresh-tokens.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateGenreDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "tmdbId": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateMovieDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1
          },
          "original_title": {
            "type": "string",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "video": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "vote_average": {
            "type": "number",
            "minimum": 0,
            "maximum": 10,
            "nullable": true
          },
          "imdbRate": {
            "type": "number",
            "minimum": 0,
            "maximum": 10,
            "nullable": true
          },
          "imdbLink": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "isWatchedTogether": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateMovieDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1
          },
          "original_title": {
            "type": "string",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "video": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "popularity": {
            "type": "number",
            "nullable": true
          },
          "original_language": {
            "type": "string",
            "nullable": true
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "backdrop_path": {
            "type": "string",
            "nullable": true
          },
          "adult": {
            "type": "boolean",
            "nullable": true
          },
          "vote_average": {
            "type": "number",
            "minimum": 0,
            "maximum": 10,
            "nullable": true
          },
          "vote_count": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "genres": {
            "maxItems": 100,
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieGenreDto"
            }
          },
          "imdbRate": {
            "type": "number",
            "minimum": 0,
            "maximum": 10,
            "nullable": true
          },
          "imdbLink": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "dbId": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "isWatchedTogether": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": [
          "title",
          "dbId"
        ],
        "additionalProperties": false
      },
...
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "tmdbId": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "tmdbId"
        ]
      },
      "GenreResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "tmdbId": {
            "type": "integer"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "tmdbId",
          "apiVersion"
        ]
      },
      "GenreListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "apiVersion"
        ]
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "itemsPerPage": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer"
          },
          "currentPage": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": [
          "itemsPerPage",
          "totalItems",
          "currentPage",
          "totalPages"
        ]
      },
      "PaginationLinks": {
        "type": "object",
        "properties": {
          "first": {
            "type": "string"
          },
          "previous": {
            "type": "string"
          },
          "next": {
            "type": "string"
          },
          "last": {
            "type": "string"
          },
          "current": {
            "type": "string"
          }
        },
        "required": [
          "first",
          "previous",
          "next",
          "last",
          "current"
        ]
      },
...
      "MovieResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dbId": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "vote_average": {
            "type": "number",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "original_title": {
            "type": "string",
            "nullable": true
          },
          "video": {
            "type": "string",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "imdbRate": {
            "type": "number",
            "nullable": true
          },
          "imdbLink": {
            "type": "string",
            "nullable": true
          },
          "isWatchedTogether": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "addedBy": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PublicUser"
              },
              {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "dbId",
          "title",
          "poster_path",
          "vote_average",
          "overview",
          "original_title",
          "video",
          "release_date",
          "imdbRate",
          "imdbLink",
          "isWatchedTogether",
          "createdAt",
          "updatedAt",
          "addedBy",
          "genres",
          "apiVersion"
        ]
      },
      "MovieListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Movie"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "meta",
          "links",
          "apiVersion"
        ]
      },
      "CreatedMovieResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dbId": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "vote_average": {
            "type": "number",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "original_title": {
            "type": "string",
            "nullable": true
          },
          "video": {
            "type": "string",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "imdbRate": {
            "type": "number",
            "nullable": true
          },
          "imdbLink": {
            "type": "string",
            "nullable": true
          },
          "isWatchedTogether": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "addedBy": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              }
            },
            "required": [
              "id"
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "dbId",
          "title",
          "isWatchedTogether",
          "addedBy",
          "genres",
          "createdAt",
          "updatedAt",
          "apiVersion"
        ]
      },
      "UpdatedMovieResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dbId": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "imdbRate": {
            "type": "number",
            "nullable": true
          },
          "imdbLink": {
            "type": "string",
            "nullable": true
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "isWatchedTogether": {
            "type": "boolean"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "addedBy": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": true
              },
              "avatar": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "username",
              "avatar"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "dbId",
          "title",
          "release_date",
          "imdbRate",
          "imdbLink",
          "poster_path",
          "overview",
          "isWatchedTogether",
          "genres",
          "addedBy",
          "createdAt",
          "updatedAt",
          "apiVersion"
        ]
      },
...
      "EmptyResponse": {
        "type": "object",
        "properties": {
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "apiVersion"
        ]
      },
...
      "TmdbGenre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "TmdbGenresResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbGenre"
            }
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "apiVersion"
        ]
      },
...
      "TmdbSearchResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "total_results": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbSearchMovie"
            }
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "page",
          "total_results",
          "total_pages",
          "results",
          "apiVersion"
        ]
      },
      "TmdbMovieResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "original_title": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "release_date": {
            "type": "string"
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "backdrop_path": {
            "type": "string",
            "nullable": true
          },
          "adult": {
            "type": "boolean"
          },
          "popularity": {
            "type": "number"
          },
          "vote_average": {
            "type": "number"
          },
          "vote_count": {
            "type": "integer"
          },
          "original_language": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbGenre"
            }
          },
          "video": {
            "type": "boolean"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "genres",
          "apiVersion"
        ],
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}